+++ applications/fishing-map/src/features/user/User.tsx
@@ -1,28 +1,31 @@
-import React, { useCallback, useEffect, useState } from 'react'
+import React, { Fragment, useCallback, useEffect, useState } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import { useTranslation } from 'react-i18next'
-import Link from 'redux-first-router-link'
 import Button from '@globalfishingwatch/ui-components/dist/button'
 import Spinner from '@globalfishingwatch/ui-components/dist/spinner'
-import IconButton from '@globalfishingwatch/ui-components/dist/icon-button'
 import GFWAPI from '@globalfishingwatch/api-client'
 import { AsyncReducerStatus } from 'types'
 import {
   fetchWorkspacesThunk,
   selectWorkspaceListStatus,
 } from 'features/workspaces-list/workspaces-list.slice'
-import { HOME, WORKSPACE } from 'routes/routes'
-import { WorkspaceCategories } from 'data/workspaces'
+import { HOME } from 'routes/routes'
 import { updateLocation } from 'routes/routes.actions'
+import { selectUserDatasets } from 'features/datasets/datasets.selectors'
+import { fetchDatasetsByIdsThunk, selectDatasetsStatus } from 'features/datasets/datasets.slice'
 import styles from './User.module.css'
 import { fetchUserThunk, GUEST_USER_TYPE, logoutUserThunk, selectUserData } from './user.slice'
 import { isUserLogged, selectUserWorkspaces } from './user.selectors'
+import UserWorkspaces from './UserWorkspaces'
+import UserDatasets from './UserDatasets'
 
 function User() {
   const { t } = useTranslation()
   const dispatch = useDispatch()
   const userLogged = useSelector(isUserLogged)
   const userData = useSelector(selectUserData)
+  const datasets = useSelector(selectUserDatasets)
+  const datasetsStatus = useSelector(selectDatasetsStatus)
   const workspaces = useSelector(selectUserWorkspaces)
   const workspacesStatus = useSelector(selectWorkspaceListStatus)
   const [logoutLoading, setLogoutLoading] = useState(false)
@@ -30,6 +33,7 @@ function User() {
   useEffect(() => {
     if (userLogged && userData?.id) {
       dispatch(fetchWorkspacesThunk({ userId: userData?.id }))
+      dispatch(fetchDatasetsByIdsThunk([]))
     }
   }, [dispatch, userData?.id, userLogged])
 
@@ -77,35 +81,15 @@ function User() {
       {/* <div className={styles.views}>
         <label>Your private views</label>
       </div> */}
-      <div className={styles.views}>
-        <label>Your latest saved views</label>
-        {workspacesStatus === AsyncReducerStatus.Loading ? (
-          <Spinner size="small" />
-        ) : (
-          <ul>
-            {workspaces?.map((workspace) => {
-              return (
-                <li className={styles.workspace} key={workspace.id}>
-                  <Link
-                    className={styles.workspaceLink}
-                    to={{
-                      type: WORKSPACE,
-                      payload: {
-                        category: workspace.category || WorkspaceCategories.FishingActivity,
-                        workspaceId: workspace.id,
-                      },
-                      query: {},
-                    }}
-                  >
-                    <span className={styles.workspaceTitle}>{workspace.name}</span>
-                    <IconButton icon="arrow-right" />
-                  </Link>
-                </li>
-              )
-            })}
-          </ul>
-        )}
-      </div>
+      {workspacesStatus !== AsyncReducerStatus.Finished ||
+      datasetsStatus !== AsyncReducerStatus.Finished ? (
+        <Spinner size="small" className={styles.spinner} />
+      ) : (
+        <Fragment>
+          <UserWorkspaces workspaces={workspaces} />
+          <UserDatasets datasets={datasets} />
+        </Fragment>
+      )}
     </div>
   )
 }