+++ packages/api-types/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @globalfishingwatch/api-types
 
+## 2.7.1
+
+### Patch Changes
+
+- fix main api-types package.json
+
+## 2.7.0
+
+### Minor Changes
+
+- include datasetType enum
+
 ## 2.6.2
 
 ### Patch Changes